requirements.txt
python-binance
websocket-client
python-dotenv
redis
pymongo
requests
numpy
pandas
scipy